+++ src/acceleration/cuda_backend.cpp
@@ -148,7 +148,7 @@ bool validateDistanceOutputs(const std::vector<float>& distances, bool useL2, st
 
 bool validateTopKOutputs(const std::vector<int>& topkIndices, const std::vector<float>& topkDistances, size_t numVectors,
                          bool useL2, std::string& validationError) {
-    if (static_cast<int>(topkIndices.size()) != static_cast<int>(topkDistances.size())) {
+    if (topkIndices.size() != topkDistances.size()) {
         validationError = "Top-K output size mismatch between indices and distances";
         return false;
     }
@@ -752,7 +752,7 @@ CUDAGraphEntry *CUDAGraphCache::get(const QueryShape &shape) noexcept {
 
 CUDAGraphEntry &CUDAGraphCache::put(const QueryShape &shape, CUDAGraphEntry entry) {
     // Only evict if this is truly a new key (not a replacement)
-    if (static_cast<int>(entries_.size()) >= kMaxEntries && entries_.count(shape) == 0) {
+    if (entries_.size() >= kMaxEntries && entries_.count(shape) == 0) {
         evictLRU();
     }
     entry.lastAccess = ++clock_;
@@ -1123,7 +1123,7 @@ CUDAGraphBFSEntry *CUDAGraphBFSCache::get(const GraphBFSShape &shape) noexcept {
 }
 
 CUDAGraphBFSEntry &CUDAGraphBFSCache::put(const GraphBFSShape &shape, CUDAGraphBFSEntry entry) {
-    if (static_cast<int>(entries_.size()) >= kMaxEntries && entries_.count(shape) == 0) {
+    if (entries_.size() >= kMaxEntries && entries_.count(shape) == 0) {
         evictLRU();
     }
     entry.lastAccess = ++clock_;
@@ -1208,7 +1208,7 @@ CUDAGraphSPEntry *CUDAGraphSPCache::get(const GraphSPShape &shape) noexcept {
 }
 
 CUDAGraphSPEntry &CUDAGraphSPCache::put(const GraphSPShape &shape, CUDAGraphSPEntry entry) {
-    if (static_cast<int>(entries_.size()) >= kMaxEntries && entries_.count(shape) == 0) {
+    if (entries_.size() >= kMaxEntries && entries_.count(shape) == 0) {
         evictLRU();
     }
     entry.lastAccess = ++clock_;
